scripts/requirements.txt
python-frontmatter>=0.5.0
PyYAML>=6.0
geojson>=2.5.0
beautifulsoup4>=4.12.0
requests>=2.31.0
# Force cache invalidation
